SystemUI: Tile: ADB over Network: lockscreen and global
Disable the ADB over Network when we are on the lockscreen
with security enforced AND if ADB over Network is not enabled.

Disable ADB over Network always if ADB is disabled.

This replace state.visible unsupported since 7.0 and
avoid any confusion for the end user because the tile is
still visible but disabled.

+ Migration from Settings.Secure to Settings.Global