BluetoothManagerService: workaround about bluetooth not always able to be toggle ON
If bluetooth name and address are not present,
message MESSAGE_GET_NAME_AND_ADDRESS can be received before MESSAGE_BLUETOOTH_SERVICE_CONNECTED
in which case mGetNameAddressOnly is set to true and then when MESSAGE_BLUETOOTH_SERVICE_CONNECTED is received
listeners are not notified that bluetooth is ON. and then when MESSAGE_BLUETOOTH_SERVICE_CONNECTED is received

This is a workaround and a better logic should be written.

For our needs this is enough.

Clear name and address for test purpose can be done in this way:
adb root
adb shell
settings delete secure bluetooth_name
settings delete secure bluetooth_address
reboot

thanks @Millosr